pyproject.toml
[tool.poetry]
name = "hammer-systems-test"
version = "0.1.0"
description = "Test case for Hammer Systems"
authors = ["Stanislav Meshcheriakov"]
readme = "README.md"
packages = [{include = "hammer_systems_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.4"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
